Batch AI amphibious moves.
Prior to this change, the AI would move load and move transports
one unit at a time. So moving 20 units via transports would take
40 moves: 20 x (1 load, 1 load, 1 move, 1 unload). With this change,
it's just 3 moves (when they all have the same destination):
  1. Load 20 units
  2. Move 10 transports
  3. Unload 20 units.

This makes a big difference when playing with AI since this makes
the AI moves be much faster when amphibious moves are involved.