Deprecate PlayerNameValidation methods & reduce dependency on magic lobby watcher node name.
With game listings moved to Http-Server, the updated system design
will not rely on naming conventions to identify automated hosts.

With this in mind, validation between server side and client side
becomes unified and can be the same logic.

The validation logic between PlayerName and PlayerNameValidation
then also becomes identical, we should stop using PlayerNameValidation
in favor of PlayerName. PlayerName can be accessed from multiple
projects, hence that is the favored copy.