Infrastructure: status check scripts
Add a few scripts to do status or health checks on prerelease.
These are meant to be helper scripts so that we can document
post-deployment expectations and not run these by hand.

The status checks would be better replaced for monitoring checks,
in the short term these are a cheap way to document and automate
health checks.