Fix KeyBindings firing while chatting
While a user is typing in a text component, we do not want keybindings
to fire. For example, pressing space key skips current unit, typing
a few words would skip a few units.

To fix this we can set a keyboard focus listener to fire when keyboard
focus changes. We then check if focus is on a text component, if so
we disable key binding actions. When focus is off of the text component,
we re-enable the key binding actions.

Addresses: https://github.com/triplea-game/triplea/issues/5286