PR Auto-Correct
Assuming maintainer edits is enabled, this update will have the builder-bot
run spotlessApply and push any updates back to the PR branch. Also, builder-bot
will look for any unnecessary lombok '@Log' tags and remove them if there
is no logging in the respective class file.

This opens up a framework for additional auto-corrections as we can think
of them.