Set min-size of action panel to 240px
Allows enough space for unit-scroller and unit-move-history
to typically not be cut-off on the right hand side.

Additionally, if there is not enough room for unit scroller,
re-sizing will yield a min-size, so you will not be able to get
the action panel back to it's original (cut-off) width.

This update sets the min width of the action panel to match the
effective min width.