Websocket fixes for HTTPS & Disconnect Handling
1. Fix disconnect window appearing when client closes the websocket connection.
This is done by splitting the concept of a connection lost listener and connection
closed listener. The closed listener is executed every time we close connection,
where client or server initiated. The connection lost listener is executed only
when the connection is closed without the client calling the close method. This
is done at the implementation level by removing connection lost listeners when
'close' is called.

2. Fix missing lobby shutdown on chat connection closed. Previously, if server
disconnected, the lobby would close but the game-headed application would continue
to run. With a shutdown call, the game-client lobby window now closes completely
and its background process, and the main game-headed initial screen is shown.

3. Fix protocol for secure websocket. Our current http client is able to convert
a http connection to websocket (ws) connection without issue. Converting an
'https' connection to 'wss' is not automatic and we must use the correct protocol
to begin with. We use 'http' for local development, 'https' for deployed server.
In this update, if we see a 'https' protocol, when establishing a websocket connection
we'll then update the protocol to be a 'wss' (secure websocket).