Add automatic prerelease deployment to travis
- Sets up Travis to do a post-release task to deploy prerelease.
This is done in short by running ansible for the prerelease hosts.

- Update README with some ansible quick overview notes and typical
commands for running deployments.

- Update 'run_deployment' script to be more generic and allow
passing of arbitrary args to ansible. This way we can choose to
use the '-d' flag on local but not in travis. '-d' shows file
diffs of what will change on remote servers, very useful for
development and debugging, but would otherwise be dangerous
as any passwords that are written to files on remote servers
would show up in that diff.