Move LobbyWatcherThread ownership to ServerModel, prep for ServerPoller
Change summary:
- Move LobbyWatcherThread ownership from SetupPanel to ServerModel
- Remove unused moderator socket remote actions
- Extract headless 'LaunchAction' anonymous inner class to its own class

ServerPoller will be a background thread on hosts to check for recently
banned players. This will allow moderators to kick/ban players from hosted
games.

To get there we need to have access to both ServerMessengers and the API
key used in LobbyWatcher-Thread. The place where we create lobby-watcher-thread
will be the same where we will create the upcoming server-poller.