Fix Node null address when using unresolved IP addresses
`InetSocketAddress.getAddress()` returns null if the address being
returned is unresolved. An easy way to create an `InetSocketAddress`
is to use: `InetSocketAddress.createUnresolved("localhost")`, though
doing this will cause the IP address passed in to be converted to
null and not stored.  IE:  `Node.getIpAddress()` will return null
instead of "localhost"

To fix this, we use a different InetSocketAddress to get the host
address as a string and we then convert that string to an InetAddress.