Game hosts banned user listener and shutdown
Introduce concept of 'remote actions', these are moderator actions
sent to the server. So far that is 'ban player' and 'shutdown' host.
This works by setting up a websocket between game hosts and server.
Via HTTP RemoteActionsClient, moderator will send shutdown messages
which are then added to the remote actions queue. The remote
actions queue sends a message to the appropriate connected websocket
connections.

Similar is done for banning, existing ban paths on the server now
emit a ban message to the remote actions event queue which notifies
the websocket listeners.

__Message Handling__
* Game hosts when they receive a shutdown message will shutdown.
* Game hosts when they receive a ban message will have one of three
behaviors:
- banned player is not connected, no-op
- banned player is connected, they are disconnected
- banned player is the host, the server messenger/host shuts down

__Future TODO__
- need to prevent banned users from re-joining, need to add
ban check when users join a host. The client endpoint is set up
for this.

__Change List__
- Pattern for websocket messages is updated quite a bit. Server
and client message envelope is kept generic and we split
the message factories between chat and remote actions. We can
emulate this pattern for additional websocket messages by
adding another factory.

- Getting payload from a server envelope is simplified, the type
is specified when getting the payload.

- Simplify generic websocket testing by decoupling message
listening from construction. We do this by making the messaage
listener a setter of generic websocket.

- Add RemoteActionsClient, http client for sending remote
actions (shutdown action) and doing ban check (for the future todo)

- Add RemoteActionsWebocket, this serves as a listener, fired
up by game hosts, listens for ban/shutdown messages and executes
necessary actions.

- Add backend remote actions event queue to issue messages to the
websocket listeners and wire message sources to the event queue.