Extract common websocket listener code to base classes
- Adding more message type enums leads to redundant code to 'ChatServerMessageType'. This
code was abstracted to a common interface 'WebsocketMessageType'.
- Updated RemoteActionsMessageType to use the common interface 'WebsocketMessageType'.
- The websocket listener itself became also very boilerplate, common code
was extracted to a common base class 'WebsocketListener'

Benefits:
- if adding a new listener websocket, you can extend WebsocketListener and just need
to define one abstract method where we plugin a call to the message type enum 'valueOf'
- message type enum becomes very boilerplate and is easily tested. Once the listener
and message type are defined and tested, then focus can be spent on the actual listeners
and no more code is needed for the actual websocket communication (though care should be
taken to make sure we close our websocket connection).