Wire moderator shutdown host option (#5678)
Adds a game table right click menu for moderators to send a shutdown
request to a game host. Note, the shutdown is by IP, all games
hosted on that IP will receive the shutdown signal.

The backend parts of this update are in place, just needed to
add the front-end components.