Refactor movable units filtering to a helper class.
This decouples the UI code in MovePanel from the logic of
filtering which units could move on a given route.

No functionality changes. This refactoring will allow us to
more easily unit test the functionality as we make changes
to it, which I plan to do in a follow-up PR.