Fix run_ansible_vagrant script to build jars
1. Invoke all functions from a 'main' method and invoke the main method
at the bottom of the file. This keeps the logic flow at the top of the file.

2. The jar creation commands were not very correct, with this update
we will use gradle to do a no-op if jars are already created rather than
checking for it in shell script, and secondly we'll create jars of only
the suprojects that we need.