Automate localhost install of cert from vagrant machine
Automates installation of NGINX SSL certificate to localhost (ansible control machine)
when deploying to a vagrant VM. Assumes localhost is linux-flavored. If running on windows OS,
will likely need to do a linux VM to simulate the control machine, and then can use vagrant
from within the VM to simulate a target machine.

Change summary:
1. Use vagrant group_vars to add override for vagrant env to place certs in '/vagrant/.vagrant/'
The '/vagrant/' folder is a special mount on the virtual machine that maps to the local machine:
  'triplea/infrastructure/'. Hence '/vagrant/.vagrant/' maps to 'triplea/infrastructure/.vagrant'

2. Use run_ansible_vagrant script to install generated cert on localhost:
 A) symlink the nginx cert from 'infrastructure/.vagrant/(cert-file)' to '/usr/local/share/ca-certificates/'
 B) when done with ansible deployment, run 'update-ca-certificates -f' to refresh certificates
    and pick up any new or updated certs in '/usr/local/share/ca-certificates/'

3. Update cert installation step to run if either cert file does not exist.