Add ansible playbook tags (#5719)
Add tags so we can install 'nginx', 'bot' or 'http_server'
on their own with, eg: `./run_ansible_vagrant -t nginx`

The apt-update command needs to be run (usually) before apt install,
hence the apt-update role in playbook is tagged with each role. Arguably
we could tag the java role as well, this tagging more or less assumes
we're doing an update and that we've already run the deployment
automation and that we're iterating on a specific role to fine-tune
the deployment for that role.