Refactor movable units filtering to a helper class. (#5708)
No behavior changes. This is split off from my other PR to
speed this code up, but decided to split out these clean up
changes to make the PR that changes behavior smaller to
review.

* Fix incorrect units loaded when loading a transport.

This regressed in #5416, because unitsToTransports was only
computed once after that commit, instead of updated each time
the unit list was filtered.

This fixes #5706.

* Refactor movable units filtering to a helper class.

This decouples the UI code in MovePanel from the logic of
filtering which units could move on a given route.

No functionality changes. This refactoring will allow us to
more easily unit test the functionality as we make changes
to it, which I plan to do in a follow-up PR.