Fix bad lag when trying to move lots of units from a territory. (#5692)
* Cleanup and refactor movement code in MovePanel.

No behavior changes. This is split off from my other PR to
speed this code up, but decided to split out these clean up
changes to make the PR that changes behavior smaller to
review.

* fix typo

* Address comments: Don't use Tuple & other clean ups.

* restore isEmpty() checks

* Fix incorrect units loaded when loading a transport.

This regressed in #5416, because unitsToTransports was only
computed once after that commit, instead of updated each time
the unit list was filtered.

This fixes #5706.

* Refactor movable units filtering to a helper class.

This decouples the UI code in MovePanel from the logic of
filtering which units could move on a given route.

No functionality changes. This refactoring will allow us to
more easily unit test the functionality as we make changes
to it, which I plan to do in a follow-up PR.

* Attempt to address codeclimate warnings.

* Add basic unit test.

* Improve test.

* Auto-Formatting

* Restore missing line.

* Auto-Formatting

* makeResult() -> Result ctor

* Add a few more basic unit tests.

* Check for best.isEmpty() instead of units.isEmpty().

* final

* Fix typo

* Small simplification and fix compile error

* Fix test and simplify MoveValidator.getMustMoveWith() params.

* Make a helper function to reduce filterUnitsThatCanMove() complexity.

* Remove gamedata param.

* .

* Use an enum for the status.

* default case to satisfy pmd

* Remove test ctor

* Rebase move optimizations onto refactor.

* Fix comment.

* Auto-Formatting

* Add a unit test for mech infantry.

* Make some helper methods.

* Clean up more.

* Restore UnitComparatorTest from earlier patchset.

* Set warning message on warning.

* Rebase on master.

* Comment fix.