Improve movable unit filtering for transport loads.
Previously, a move of 3 infantry and 2 tanks to one transport
would not result in the algorithm choosing 1 tank and 1 infantry
to load - instead, it loaded just one tank.

This is because it sorted the list as {inf, inf, inf, tank, tank}
and tried all tail sublists of that list to see what fits.

This change fixes that by just using the mapTransports() algorithm
to choose the units that can be loaded and adds unit tests.