Fix websocket send race-condition, make send wait and block. (#5750)
* Fix websocket send race-condition, make send wait and block.

Summary:
- remove single threaded threadpool for websocket connect and send
- make connect be non-blocking
- have send wait for connection to be open (now blocking)

Problem description:
Observed preconditions failures in 'send' for connection closed.
The previous flow relied on single-threaded operations to ensure
we opened a connection and then sent data over an open connection.

This was ordered by queuing operations onto a single-threaded thread
pool. The suspected problem is that (A) the connect thread spawns
a new thread (B) and waits for it to be finished before we spawn
a new thread (C) to do the send. If (C) is spawned before (B), we
hit a race condition and attempt to send befor the websocket connection
is open.