Non-blocking websocket send
First, do a blocking connect on a new thread and add a timeout to it. This way
if we fail to connect we will call an error handler to notify the connection
failed. Before, if we never connected, there was never a timeout nor notification.

Second, add a send queue to websocket sender. On send, check if connection is open,
if not open, add the message to send to the send queue. On connection open,
we will flush the send queue. We have a sync block to make sure that
before we check if the connection is open, if we are flushing messages, that
we cannot queue messages. This is to avoid a race condition where after
we check if the connection is open, if it opens then and we then
flush the send queue and clear it, we want to avoid adding any new messages
to the send queue.

Change summary:
- Blocking connect on a new thread with a timeout invokes error handler
- Wire error handler to websocket sender
- Make send non-blocking by using a send queue
- Minor: add lombok getter/setter for test and code cleanup