Fix focus lost when typing on PBEM screen.
This was caused by MainPanel having two types of observers and
a bug introduced in 895115c that caused the wrong one to be called.

This change changes reverts that error and renames the two
mechanisms to disambiguate them - one is about setting a new set
up panel and the other is about updating the button to start the
game.