Remap flag hotkey 'L' to ctrl+(1-3), remove show flags button
This update binds:
ctrl+1 = no flags
ctrl+2 = small flags
ctrl+3 = large flags

We also remove the 'L' hotkey listener that toggled through flag display
modes. With the hotkeys now labeled in the game view menu, and no more
'L' hotkey that was otherwise unlabelled, we can remove the not very
popular flag mode button that was in the unit scroller.