Add 2.0 to 1.9 draw_mode 'enum not found' compatibility hack (#5858)
The problem being addressed is specifically when loading a 2.0 game,
then loading a 1.9 game will crash.

This update addresses that problem where a 1.9 game tries to load an
enum from a setting key that was set in later versions to a value that
is not contained in the 1.9 enum. This patch clears out the preferences
key which allows 1.9 to behave properly.