Remove output install4j buffering
Error messages are being printed currently from install4j (perhaps they
are being printed now to stdout instead of stderr). This update
removes functionality to cache error messages to a temporary
file for the purpose of suppressing those error messages. It was previously
important as the error message volume was sufficient to run out of
console buffer on travis resulting in truncated messages.