Add test email button tooltip and mark 'cc' label as required
1. Send test email button is disabled without a tooltip. This update
adds a tooltip to explain that both 'to' and 'cc' email fields
need to be entered first.

2. 'cc' label does not show up as red text when empty even though
it is required by dice server to have a 'cc' email address. This
update makes the 'cc' field show as red when empty and does not
enable the 'test email' button until this field is non-empty
with a valid email address.