Do not double handle map not found exception
Currently if a map not found exception is raised when loading a game
we both prompt a user if they would like to download the map and
secondly we notify that the game failed to start and provide
an option to upload an error report.

This update removes the latter handling so that when a map is not
found, we simply notify and prompt the user to download the missing
map.