Compute units to render in unit scroller before adding a task to EDT
Simply reduces the amount of computation done on EDT. EDT can take
a substantial amount of time to render, results computed on such
a late scheduled thread may no longer be concurrent. Additionally,
iterating on lists on a new thread (EDT) that can be modified from
the main thread opens a possibility of ConcurrentModificationExceptions.

This change simply moves rendering computations to be earlier so that EDT
usese the results of these computations instead of computing them itself.