Avoid unnecessary unit scroller rendering
If the move panel is not visible, then we do not need to render the unit
scroller and we should avoid any related computation. The unit scroller
is tied into game events, notably phase changes so it can update itself.
Missing from these listeners was the concept to check if the rendering
would have been necessary.

This update wires a supplier function to the unit scroller that can
be used to determine if the parent panel is even visible. If the parent
panel is not visible, then rendering computations are skipped.