Add username and password fields to PBF selection screen (#5980)
* Add username and password fields to PBF selection screen

The two fields are required in order to do a test posting or
to start a PBF game. This update places the two required fields
on the same screen where we have the 'play' button.

This update adds similar 'highlight field label red if invalid'
for the username and password field.

Of note, to avoid circular event callbacks between view and
view-model, the view updates its own state after firing an
event on the view-model. For example, the username field
sets the view-model username value, the view then queries
and updates to see if any labels or buttons are updated and
enabled. The view model could invoke the view to update all
fields, this can then trigger another callback though to the
view when fields are updated and action listeners fire to
update the view model which can then again trigger a view
refresh and so-on.

* Use non-guava for predicate

* Update game-core/src/main/java/games/strategy/engine/framework/startup/ui/posted/game/pbf/ForumPosterEditorViewModel.java

Co-Authored-By: RoiEX <8350879+RoiEXLab@users.noreply.github.com>

Co-authored-by: RoiEX <8350879+RoiEXLab@users.noreply.github.com>