Fix password too short error message when creating an account
Previous password validator assumed we were looking at a hashed value,
now we are looking at a plaintext value. The server validation of
length is still needed as there is no guarantee someone will use
our game-client when creating accounts.

This fix updates password validation on account creation to simply
check that the password is a minimum length.