Update DocumentListenerBuilder to have a buffering callback API
Forward looking change to add functionality to DocumentListenerBuilder
that will buffer text listener callbackss.

The text update callback will be scheduled to be executed only after
an initial delay after the first text update event. But, if another
text update comes along, the previously scheduled callback is canceled
and a new one with delay is scheduled.

This allows for essentially (more or less) one callback to be done
after a user is done typing text rather than rapid firing many callback
events.

Usages of the previous non-buffered API will be replaced with the new API
in upcoming PRs