Rewrite PbemEditor to prompt for email provider and settings
Pbem has as required settings an email server that are specified in
preferences, a different screen. This causes a problem where the
'play' button cannot be enabled on PBEM by simply filling in all fields
on the PBEM screen.

This update adds a few of these settings from the preferences screen
to PBEM, notably the 'email provider' setting. The email provider
setting starts out disabled, in which case the play button is now enabled.
This replicates previous behavior of the 1.9.0.0.7xxx release where
PBEM server was optional when playing a PBEM game.

Still to do after this update is to add 'username' and 'password' as
fields on the PBEM screen, after that point all required fields
can be specified on the screen that needs them.