Remove ctrl-z hotkey
The hotkey collapses the right hand action bar (maps+tabs).
The hotkey is being removed to avoid a bug where pressing
the hotkey twice in a row resulted in a error. Secondarily,
the hotkey did not have good discoverability, was not highly
used, and there is a UI control on the side-bar to close the
sidebar.

Addresses: https://github.com/triplea-game/triplea/issues/5847