Pbem settings (#6026)
* Update DocumentListenerBuilder to have a buffering callback API

Forward looking change to add functionality to DocumentListenerBuilder
that will buffer text listener callbackss.

The text update callback will be scheduled to be executed only after
an initial delay after the first text update event. But, if another
text update comes along, the previously scheduled callback is canceled
and a new one with delay is scheduled.

This allows for essentially (more or less) one callback to be done
after a user is done typing text rather than rapid firing many callback
events.

Usages of the previous non-buffered API will be replaced with the new API
in upcoming PRs

* Rewrite PbemEditor to prompt for email provider and settings

Pbem has as required settings an email server that are specified in
preferences, a different screen. This causes a problem where the
'play' button cannot be enabled on PBEM by simply filling in all fields
on the PBEM screen.

This update adds a few of these settings from the preferences screen
to PBEM, notably the 'email provider' setting. The email provider
setting starts out disabled, in which case the play button is now enabled.
This replicates previous behavior of the 1.9.0.0.7xxx release where
PBEM server was optional when playing a PBEM game.

Still to do after this update is to add 'username' and 'password' as
fields on the PBEM screen, after that point all required fields
can be specified on the screen that needs them.

* Remove grid bag layout from dice server editor, not used

* Add clarifying comment on why text fields should not be updated to the same value (moves the carat to end of field)