Move class 'ChatParticipant' to domain-data (#6051)
Moved in anticipation of a structure split in http-client, will want to use
'chatParticipant' across additional sub-projects. It's also been an ongoing
tension of domain data having been defined in http-client, as that has
been a least common denominator for shared data, arguably data that
is more fundamental than http messages should not live there to begin with.