Combine moderator ban controllers
The first moderator ban controller was built to enable the moderator toolbox.
We had a second one come about to support the moderator right-click-ban menu
option. This update combines the two.

A bit forward looking, this update adds factory methods directly on the classes
being instantiated. A follow-up update is planned to make this pattern of
factory insantiation to be consistent.