Group backend services (#6082)
* Re-group spitfire server packages

Fixes a few package organization issues that have been lingering
and preps for design changes to server action handling.

- group moderation package together
- create a concept of 'modules', meant to represent the business domains.
  We still will have strongly cohesive modules that contain all necessary
  components to accomplish a functionality (controller+service)
- move web.socket connection components to a new package

* Combine moderator ban controllers

The first moderator ban controller was built to enable the moderator toolbox.
We had a second one come about to support the moderator right-click-ban menu
option. This update combines the two.

A bit forward looking, this update adds factory methods directly on the classes
being instantiated. A follow-up update is planned to make this pattern of
factory insantiation to be consistent.

* Update jar path to main class "ServerApplication"