Make findNearbyEnemyCapitalsAndFactories() ~30x faster. (#6086)
On great war map with all AIs, after 1700 executions, the new
version takes ~150ms in aggregate while the previous
implementation takes 4900ms.

The algorithm's functionality is kept the same with the exception
of removing the max distance constant, which was only there as a
performance optimization. It's not needed with the new algorithm.