Sort unit scroller units by air/sea/land; then power; then name
From left to right, place:
- air units
- sea units
- land units

Within each category, place stronger attack units
on the left, then break any ties by sorting by name.

We sort by attack and not defense or a max of the two
as the unit scroller is viewed during a players combat
move, the attack power is the operative value during
that time.