Rename 'station' back to 'sleep' and revert back to the sleeping moon icon
The term 'station' was to avoid 'sleep' and seems to no have stuck
very well. Sleep is a bit more self explanatory even if just marginally so.

Second, the shield icon for station was not a super great mapping to
the 'station' concept and on initial perception is not as clear as the
concept of 'sleep'.