Add default save game name
This is done by adding a gameData property that is the save game file name
that we are loading from. The property is null if we are starting a new game.

In the case wheere we are loading new game, instead of defaulting
the save game file name to the name of the save game, we instead create
 a default name with this pattern "YYYY-MM-DD-map-name.tsvg"