Remove focusable from 'alert' button
Clicking the alert button repeatedly has no real effect, a user
typically does not want to do this. After a first click, the button
gains focus. *Then* pressing space will activate the button. This
is confusing behavior as the next available unit is selected
and it seems like we are 'skipping' units with the space button
but we are really pressing 'wake-all' repeatedly.

To avoid this confusing scenario, remove focusable from 'alert'
button. This way if it is clicked, and then a user presses space,
that will activate the 'skip' button.