Fix transposed time units when banning from toolbox
The time unit from toolbox ban was transposed and was being sent
as minutes instead of as hours or days.

The fix here is two-fold, one we consolidate a duplicate ban
duration prompt to re-use the ban duration prompt available
when right-clicking player names. Using this ban prompt
we can then send the proper minute time value for ban durations.