Change status of games with zero players to be 'available'
This change updates client code to convert the status of games
with zero players from 'waiting for players' to instead 'available'.

This is done client side for ease of implementation and avoid
headless games from having to re-report a new status as players join
and leave.

The benefit to this change is sorting the list games by game status,
'waiting for players' will truly now only have games waiting for players.
This is a problem when there are many available bot servers but only
a few with players waiting. With this update the sort by game status
will offer correctly games that are 'waiting for players' vs those
that are empty.