Add validation text to change password form
This update adds validation message to the 'change password' form.
Notably before there was no messaging of what was wrong with the
passwords such that the 'okay' button was not enabled. This update
adds validation text that prints in red when password and password
confirm do not match or is too short (which should explain why
okay button is not enabled).