Specify error prone version in http server, avoid version conflict (#6178)
Error prone pulls in an older version of caffeine that conflicts
with the version we need. This problem is fixed in 'java-extras',
but if gradle dependencies are evaluated from http-server then
we can still pick in this older version.

This update applies the same fix we did for 'java-extras' and we
explicitly specify the caffeine version to use in error prone
which prevents the wrong version from being used in test-implementation.